requirements.txt
amqp==2.6.0
ariadne==0.11.0
asgiref==3.2.7
attrs==19.3.0
awesome-slugify==1.6.5
beautifulsoup4==4.7.1
billiard==3.6.3.0
bleach==3.1.5
boto3==1.14.0
botocore==1.17.0
cachetools==4.1.1
celery==4.4.5
certifi==2020.4.5.2
cffi==1.14.4
chardet==3.0.4
Collectfast==2.2.0
cryptography==2.9.2
dbbackup-tools==0.1.5
defusedxml==0.6.0
Django==3.0.7
django-background-tasks==1.2.5
django-celery-beat==2.1.0
django-compat==1.0.15
django-cprofile-middleware==1.0.5
django-dbbackup==3.3.0
django-debug-toolbar==1.11
django-djconfig==0.10.0
django-environ==0.4.5
django-haystack==3.0
django-htmlmin==0.11.0
django-infinite-scroll-pagination==1.1.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-pwa==1.0.9
django-rest-framework==0.1.0
django-simple-sso==0.14.1
django-smart-selects==1.5.9
django-spirit==0.12.1
django-storages==1.9.1
django-timezone-field==4.0
django-user-agents==0.4.0
djangorestframework==3.11.0
docutils==0.15.2
ecdsa==0.14.1
enum-compat==0.0.3
et-xmlfile==1.0.1
Faker==5.0.2
fastdiff==0.2.0
future==0.18.2
google-api-core==1.22.2
google-api-python-client==1.7.11
google-auth==1.21.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-core==1.4.1
google-cloud-storage==1.31.0
google-cloud-vision==1.0.0
google-crc32c==1.0.0
google-pasta==0.2.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
grapheme==0.6.0
graphql-core==3.0.5
grpcio==1.32.0
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.18.1
idna==2.9
importlib-metadata==1.6.1
intuit-oauth==1.2.4
itsdangerous==1.1.0
jdcal==1.4.1
jmespath==0.10.0
kombu==4.6.10
Markdown==2.6.11
mistune==0.8.4
more-itertools==8.3.0
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.3
packaging==20.4
pandas==1.0.4
Pillow==8.0.1
pkg-resources==0.0.0
pluggy==0.13.1
protobuf==3.14.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.2
pytest-django==3.9.0
pytest-mock==3.1.0
python-arango==5.4.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
python-jose==3.2.0
python-magic==0.4.18
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.2
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.10.14
rsa==4.6
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
snapshottest==0.5.1
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
starlette==0.13.4
termcolor==1.1.0
text-unidecode==1.3
typing-extensions==3.7.4.3
ua-parser==0.10.0
Unidecode==0.4.21
uritemplate==3.0.1
urllib3==1.25.9
user-agents==2.2.0
vine==1.3.0
wasmer==0.4.1
wcwidth==0.1.9
webencodings==0.5.1
webservices==0.7
Whoosh==2.7.4
zipp==3.1.0
